move seed.go functions to proper files
